+++ .idea/deploymentTargetDropDown.xml
@@ -7,11 +7,11 @@
         <deviceKey>
           <Key>
             <type value="VIRTUAL_DEVICE_PATH" />
-            <value value="C:\Users\amber\.android\avd\Pixel_4_API_30.avd" />
+            <value value="C:\Users\amber\.android\avd\Huawei_P20_Lite_API_30.avd" />
           </Key>
         </deviceKey>
       </Target>
     </targetSelectedWithDropDown>
-    <timeTargetWasSelectedWithDropDown value="2022-09-27T09:40:57.725240200Z" />
+    <timeTargetWasSelectedWithDropDown value="2022-09-27T15:28:24.590390300Z" />
   </component>
 </project>
