+++ .idea/misc.xml
@@ -14,7 +14,10 @@
         <entry key="..\:/Users/Storm/AndroidStudioProjects/Tour-IT/app/src/main/res/menu/bottom_nav_bar.xml" value="0.371875" />
         <entry key="..\:/Users/Storm/AndroidStudioProjects/Tour-IT/app/src/main/res/menu/main_nav_bar.xml" value="0.371875" />
         <entry key="..\:/Users/amber/OneDrive/University/BCAD3/Semester 2/Open Source Coding (Intermediate)/POE/Android Studio/app/src/main/res/drawable-v24/dialogue_logout_bacground.xml" value="0.3135" />
+        <entry key="..\:/Users/amber/OneDrive/University/BCAD3/Semester 2/Open Source Coding (Intermediate)/POE/Android Studio/app/src/main/res/drawable/background_btn_blue_drawable.xml" value="0.3135" />
         <entry key="..\:/Users/amber/OneDrive/University/BCAD3/Semester 2/Open Source Coding (Intermediate)/POE/Android Studio/app/src/main/res/drawable/background_selector_2.xml" value="0.1435" />
+        <entry key="..\:/Users/amber/OneDrive/University/BCAD3/Semester 2/Open Source Coding (Intermediate)/POE/Android Studio/app/src/main/res/drawable/background_text_white_drawable.xml" value="0.3135" />
+        <entry key="..\:/Users/amber/OneDrive/University/BCAD3/Semester 2/Open Source Coding (Intermediate)/POE/Android Studio/app/src/main/res/drawable/blue_search_backdrop.xml" value="0.3135" />
         <entry key="..\:/Users/amber/OneDrive/University/BCAD3/Semester 2/Open Source Coding (Intermediate)/POE/Android Studio/app/src/main/res/drawable/btn_blue_rounded.xml" value="0.3135" />
         <entry key="..\:/Users/amber/OneDrive/University/BCAD3/Semester 2/Open Source Coding (Intermediate)/POE/Android Studio/app/src/main/res/drawable/btn_white_rounded.xml" value="0.3135" />
         <entry key="..\:/Users/amber/OneDrive/University/BCAD3/Semester 2/Open Source Coding (Intermediate)/POE/Android Studio/app/src/main/res/drawable/highlight_colour.xml" value="0.183" />
@@ -33,8 +36,9 @@
         <entry key="..\:/Users/amber/OneDrive/University/BCAD3/Semester 2/Open Source Coding (Intermediate)/POE/Android Studio/app/src/main/res/layout/bottom_drawer.xml" value="0.3625" />
         <entry key="..\:/Users/amber/OneDrive/University/BCAD3/Semester 2/Open Source Coding (Intermediate)/POE/Android Studio/app/src/main/res/layout/burger_menu.xml" value="0.3625" />
         <entry key="..\:/Users/amber/OneDrive/University/BCAD3/Semester 2/Open Source Coding (Intermediate)/POE/Android Studio/app/src/main/res/layout/dialogue_logout.xml" value="0.3625" />
-        <entry key="..\:/Users/amber/OneDrive/University/BCAD3/Semester 2/Open Source Coding (Intermediate)/POE/Android Studio/app/src/main/res/layout/fragment_account.xml" value="0.3625" />
+        <entry key="..\:/Users/amber/OneDrive/University/BCAD3/Semester 2/Open Source Coding (Intermediate)/POE/Android Studio/app/src/main/res/layout/fragment_account.xml" value="0.33" />
         <entry key="..\:/Users/amber/OneDrive/University/BCAD3/Semester 2/Open Source Coding (Intermediate)/POE/Android Studio/app/src/main/res/layout/fragment_favourites.xml" value="0.3625" />
+        <entry key="..\:/Users/amber/OneDrive/University/BCAD3/Semester 2/Open Source Coding (Intermediate)/POE/Android Studio/app/src/main/res/layout/fragment_home.xml" value="0.3625" />
         <entry key="..\:/Users/amber/OneDrive/University/BCAD3/Semester 2/Open Source Coding (Intermediate)/POE/Android Studio/app/src/main/res/layout/fragment_search.xml" value="0.31650201244054155" />
         <entry key="..\:/Users/amber/OneDrive/University/BCAD3/Semester 2/Open Source Coding (Intermediate)/POE/Android Studio/app/src/main/res/layout/fragment_settings.xml" value="0.3625" />
         <entry key="..\:/Users/amber/OneDrive/University/BCAD3/Semester 2/Open Source Coding (Intermediate)/POE/Android Studio/app/src/main/res/layout/main_burger_menu.xml" value="0.3625" />
