--- app/src/main/java/com/example/tour_it_app/FavouritesFragment.java
@@ -1,23 +1,0 @@
-package com.example.tour_it_app;
-
-import android.os.Bundle;
-
-import androidx.fragment.app.Fragment;
-
-import android.view.LayoutInflater;
-import android.view.View;
-import android.view.ViewGroup;
-
-public class FavouritesFragment extends Fragment {
-
-    public FavouritesFragment() {
-        // Required empty public constructor
-    }
-
-    @Override
-    public View onCreateView(LayoutInflater inflater, ViewGroup container,
-                             Bundle savedInstanceState) {
-        // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_favourites, container, false);
-    }
-}
