+++ app/src/main/java/com/example/tour_it_app/startup/GetStarted.java
@@ -1,4 +1,4 @@
-package com.example.tour_it_app;
+package com.example.tour_it_app.startup;
 
 import androidx.appcompat.app.AppCompatActivity;
 import androidx.appcompat.widget.AppCompatButton;
@@ -7,45 +7,45 @@ import android.content.Intent;
 import android.os.Bundle;
 import android.view.View;
 import android.widget.TextView;
-import android.widget.Toast;
 
-public class Register extends AppCompatActivity {
+import com.example.tour_it_app.R;
+
+public class GetStarted extends AppCompatActivity {
+
 
     // ---------------- * Component declaration * ------------------ //
 
-    private AppCompatButton btnSignUp;
-    private TextView btnLogin3;
+    private AppCompatButton btnGetStarted;
+    private TextView btnLogin1;
 
     // ------------------------------------------------------------ //
+
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_register);
+        setContentView(R.layout.activity_get_started);
 
-        btnSignUp = findViewById(R.id.btn_sign_up);
-        btnLogin3 = findViewById(R.id.btn_login_here_2);
+        btnGetStarted = findViewById(R.id.btn_get_started);
+        btnLogin1 = findViewById(R.id.btn_login_here_2);
 
         // --------------------------------- * Listeners * -------------------------------------- //
 
-        btnSignUp.setOnClickListener(new View.OnClickListener() {
+        btnGetStarted.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View view) {
-                //TODO: User Account creation implementation
-
-                Intent i = new Intent(Register.this, Login.class);
+                Intent i = new Intent(GetStarted.this, Register.class);
                 startActivity(i);
             }
         });
 
-        btnLogin3.setOnClickListener(new View.OnClickListener() {
+        btnLogin1.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View view) {
-               Intent i = new Intent(Register.this, Login.class);
-               startActivity(i);
+                Intent i = new Intent(GetStarted.this, Login.class);
+                startActivity(i);
             }
         });
-
 
         // -------------------------------------------------------------------------------------- //
     }
