+++ app/src/main/java/com/example/tour_it_app/startup/Register.java
@@ -1,4 +1,4 @@
-package com.example.tour_it_app;
+package com.example.tour_it_app.startup;
 
 import androidx.appcompat.app.AppCompatActivity;
 import androidx.appcompat.widget.AppCompatButton;
@@ -8,42 +8,45 @@ import android.os.Bundle;
 import android.view.View;
 import android.widget.TextView;
 
-public class GetStarted extends AppCompatActivity {
+import com.example.tour_it_app.R;
 
+public class Register extends AppCompatActivity {
 
     // ---------------- * Component declaration * ------------------ //
 
-    private AppCompatButton btnGetStarted;
-    private TextView btnLogin1;
+    private AppCompatButton btnSignUp;
+    private TextView btnLogin3;
 
     // ------------------------------------------------------------ //
-
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_get_started);
+        setContentView(R.layout.activity_register);
 
-        btnGetStarted = findViewById(R.id.btn_get_started);
-        btnLogin1 = findViewById(R.id.btn_login_here_2);
+        btnSignUp = findViewById(R.id.btn_sign_up);
+        btnLogin3 = findViewById(R.id.btn_login_here_2);
 
         // --------------------------------- * Listeners * -------------------------------------- //
 
-        btnGetStarted.setOnClickListener(new View.OnClickListener() {
+        btnSignUp.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View view) {
-                Intent i = new Intent(GetStarted.this, Register.class);
+                //TODO: User Account creation implementation
+
+                Intent i = new Intent(Register.this, Login.class);
                 startActivity(i);
             }
         });
 
-        btnLogin1.setOnClickListener(new View.OnClickListener() {
+        btnLogin3.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View view) {
-                Intent i = new Intent(GetStarted.this, Login.class);
-                startActivity(i);
+               Intent i = new Intent(Register.this, Login.class);
+               startActivity(i);
             }
         });
+
 
         // -------------------------------------------------------------------------------------- //
     }
