+++ app/src/main/res/layout/activity_login.xml
@@ -6,7 +6,7 @@
     android:layout_height="match_parent"
     android:background="@drawable/login_background"
     android:orientation="vertical"
-    tools:context=".Login" >
+    tools:context=".startup.Login" >
 
     <ImageView
         android:id="@+id/imageView3"
