+++ app/src/main/res/layout/dialogue_logout.xml
@@ -7,7 +7,7 @@
     android:orientation="vertical">
 
     <LinearLayout
-        android:layout_width="346dp"
+        android:layout_width="match_parent"
         android:layout_height="207dp"
         android:layout_gravity="center_horizontal"
         android:background="@drawable/dialogue_logout_background"
@@ -18,18 +18,18 @@
             android:layout_height="19dp" />
 
         <LinearLayout
-            android:layout_width="match_parent"
+            android:layout_width="wrap_content"
             android:layout_height="115dp"
             android:orientation="horizontal">
 
             <Space
-                android:layout_width="77dp"
+                android:layout_width="54dp"
                 android:layout_height="match_parent"
                 android:layout_weight="1" />
 
             <ImageView
                 android:id="@+id/imageView5"
-                android:layout_width="wrap_content"
+                android:layout_width="58dp"
                 android:layout_height="101dp"
                 android:layout_gravity="center_vertical"
                 android:layout_weight="1"
@@ -45,7 +45,7 @@
                 android:gravity="center_horizontal"
                 android:text="@string/logout"
                 android:textColor="@color/black"
-                android:textSize="16sp" />
+                android:textSize="15sp" />
         </LinearLayout>
 
         <LinearLayout
