+++ app/src/main/res/layout/fragment_favourites.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".FavouritesFragment">
+    tools:context=".fragments.others.FavouritesFragment">
 
     <LinearLayout
         android:layout_width="match_parent"
@@ -12,14 +12,9 @@
         android:orientation="vertical">
 
         <TextView
-            android:id="@+id/textView4"
-            android:layout_width="wrap_content"
+            android:id="@+id/textView8"
+            android:layout_width="match_parent"
             android:layout_height="wrap_content"
-            android:layout_gravity="center_horizontal"
-            android:layout_marginTop="20dp"
-            android:fontFamily="@font/poppins_semibold"
-            android:text="Favourites"
-            android:textColor="@color/black"
-            android:textSize="20sp" />
+            android:text="Favourites" />
     </LinearLayout>
 </FrameLayout>
