+++ app/src/main/res/values/colors.xml
@@ -23,4 +23,6 @@
     <color name="custom14">#2F404B</color>
     <color name="custom15">#35556C</color>
     <color name="custom16">#6391B2</color>
+    <color name="custom17">#EFEFEF</color>
+    <color name="custom18">#486D88</color>
 </resources>
